Rename hook to useUserPreference
