+++ pages/accountLists/[accountListId]/setup/finish.page.tsx
@@ -8,7 +8,7 @@ import { SetupPage } from 'src/components/Setup/SetupPage';
 import { LargeButton } from 'src/components/Setup/styledComponents';
 import { useAccountListId } from 'src/hooks/useAccountListId';
 import useGetAppSettings from 'src/hooks/useGetAppSettings';
-import { useSavedPreference } from 'src/hooks/useSavedPreference';
+import { useUserPreference } from 'src/hooks/useUserPreference';
 
 // This is the last page of the tour, and it lets users choose to go to the
 // tools page. It is always shown.
@@ -17,7 +17,7 @@ const FinishPage: React.FC = () => {
   const { appName } = useGetAppSettings();
   const accountListId = useAccountListId();
   const { push } = useRouter();
-  const [_, setSetupPosition] = useSavedPreference({
+  const [_, setSetupPosition] = useUserPreference({
     key: 'setup_position',
     defaultValue: '',
   });
